WIP - Converted the problem details exception response factory to be an exception renderer because the two concepts were so closely related.  Started dropping comments where things will need to be reworked to support configuring problem details for specific exception types.
